Refactored version check functionality out of EFB.tsx into AircraftVersionChecker class
