+++ src/instruments/src/EFB/Efb.tsx
@@ -12,9 +12,8 @@ import { usePersistentNumberProperty, usePersistentProperty } from '@instruments
 import { Battery } from 'react-bootstrap-icons';
 import { toast, ToastContainer } from 'react-toastify';
 import { distanceTo } from 'msfs-geo';
-import { PopUp } from '@shared/popup';
 import { CommitInfo, GitVersions, ReleaseInfo } from '@flybywiresim/api-client';
-import { BuildInfo } from './Utils/BuildInfo';
+import { AircraftVersionChecker } from './Utils/AircraftVersionChecker';
 import { Tooltip } from './UtilComponents/TooltipWrapper';
 import { FbwLogo } from './UtilComponents/FbwLogo';
 import { AlertModal, ModalContainer, useModals } from './UtilComponents/Modals/Modals';
@@ -120,11 +119,11 @@ const Efb = () => {
 
     // Aircraft Version check variables
     const [versionChecked, setVersionChecked] = useState(false);
-    const [buildInfo, setBuildInfo] = useState<BuildInfo | undefined>(undefined);
+    const [buildInfo, setBuildInfo] = useState<AircraftVersionChecker | undefined>(undefined);
     const [releaseInfo, setReleaseInfo] = useState<ReleaseInfo[] | undefined>(undefined);
     const [newestCommit, setNewestCommit] = useState<CommitInfo | undefined>(undefined);
     const [newestExpCommit, setNewestExpCommit] = useState<CommitInfo | undefined>(undefined);
-    const [outdatedVersionFlag, setOutdatedVersionFlag] = useSimVar('L:A32NX_OUTDATED_VERSION', 'boolean', 200);
+    const [, setOutdatedVersionFlag] = useSimVar('L:A32NX_OUTDATED_VERSION', 'boolean', 200);
 
     // Retrieves the various versions from the current aircraft and github
     const checkAircraftVersion = () => {
@@ -140,30 +139,8 @@ const Efb = () => {
             .then((releases) => setNewestExpCommit(releases))
             .catch((error) => console.error('Checking newest experimental commit failed: ', error));
 
-        BuildInfo.getBuildInfo().then((buildInfo: BuildInfo) => setBuildInfo(buildInfo))
+        AircraftVersionChecker.getBuildInfo().then((buildInfo: AircraftVersionChecker) => setBuildInfo(buildInfo))
             .catch((error) => console.error('Checking current aircraft version failed: ', error));
-    };
-
-    // Show a version info modal if the aircraft version is outdated
-    const showVersionPopup = (branchName, currentVersion, releaseVersion) => {
-        const popup = new PopUp();
-        popup.showInformation(
-            'NEW VERSION AVAILABLE',
-            `<div style="font-size: 100%; text-align: left;">
-                     You are using ${branchName} version:<br/><strong>${currentVersion}</strong><br/><br/> 
-                     Latest ${branchName} version is:<br /><strong>${releaseVersion}</strong><br/><br/>
-                     Please update your aircraft using the FlyByWire Installer.
-                     </div>`,
-            'normal',
-            () => {},
-        );
-    };
-
-    // Adds a given number of days to a given Date
-    const addDays = (date: Date, days): Date => {
-        const result = new Date(date);
-        result.setDate(date.getDate() + days);
-        return result;
     };
 
     // Called when aircraft and github version information is available/changes
@@ -175,83 +152,8 @@ const Efb = () => {
 
         // only run if we have all the information we need
         if (buildInfo && releaseInfo && newestCommit && newestExpCommit) {
-            console.log('Checking aircraft version');
-
-            // console.debug(`Current aircraft version: ${buildInfo.version}`);
-            // console.debug('Latest Released Version: ', releaseInfo[0].name);
-            // console.debug('Newest Commit: ', newestCommit.sha);
-            // console.debug('Newest Experimental Commit: ', newestExpCommit.sha);
-
-            try {
-                const versionInfo = BuildInfo.getVersionInfo(buildInfo.version);
-
-                // Set branchName to the long versions of the aircraft edition names
-                let branchName = versionInfo.branch;
-                switch (versionInfo.branch) {
-                case 'rel': branchName = 'Stable'; break;
-                case 'dev': branchName = 'Development'; break;
-                case 'exp': branchName = 'Experimental'; break;
-                default: break;
-                }
-
-                let outdated = false;
-
-                // If the users version is older than the latest release show notification
-                if (BuildInfo.versionCompare(versionInfo.version, releaseInfo[0].name) < 0) {
-                    console.log(`New version available: ${versionInfo.version} ==> ${releaseInfo[0].name}`);
-                    showVersionPopup('', versionInfo.version, releaseInfo[0].name);
-                    outdated = true;
-                } else {
-                    // If the users version is equal or newer than the latest release then check if
-                    // the edition is Development or Experimental and if the commit is older than
-                    // {maxAge} days after the latest release
-
-                    const maxAge = 3;
-                    const timestampAircraft: Date = new Date(buildInfo.built);
-
-                    if (versionInfo.branch.includes('rel')) {
-                        // Stable
-                        // console.debug('Stable version detected!');
-                    } else if ((branchName === 'Development')) {
-                        // Development
-                        // console.debug(`branch "${branchName}" version detected - checking sha: ${versionInfo.commit} against ${newestCommit.shortSha}`);
-                        if (versionInfo.commit !== newestCommit.shortSha) {
-                            if (addDays(newestCommit.timestamp, maxAge) < timestampAircraft) {
-                                const currentVersionStr = `${versionInfo.version}-${versionInfo.branch}.${versionInfo.commit} (${timestampAircraft.toUTCString()})`;
-                                const releaseVersionStr = `${versionInfo.version}-${versionInfo.branch}.${newestCommit.shortSha} (${newestCommit.timestamp.toUTCString()})`;
-                                console.log(`New commit available: ${currentVersionStr} ==> ${releaseVersionStr}`);
-                                showVersionPopup(branchName, currentVersionStr, releaseVersionStr);
-                                outdated = true;
-                            }
-                        }
-                    } else if ((branchName === 'Experimental')) {
-                        // Experimental
-                        // console.debug(`branch "${branchName}" version detected - checking sha: ${versionInfo.commit} against ${newestCommit.shortSha}`);
-                        if (versionInfo.commit !== newestExpCommit.shortSha) {
-                            if (addDays(newestExpCommit.timestamp, maxAge) < timestampAircraft) {
-                                const currentVersionStr = `${versionInfo.version}-${versionInfo.branch}.${versionInfo.commit} (timestamp: ${timestampAircraft.toUTCString()})`;
-                                const releaseVersionStr = `${versionInfo.version}-${versionInfo.branch}.${newestExpCommit.shortSha} (timestamp: ${newestExpCommit.timestamp.toUTCString()})`;
-                                console.log(`New commit available: ${currentVersionStr} ==> ${releaseVersionStr}`);
-                                showVersionPopup(branchName, currentVersionStr, releaseVersionStr);
-                                outdated = true;
-                            }
-                        }
-                    } else {
-                        // PR or any local build - no further version check
-                        // console.debug(`branch "${branchName}" version detected!`);
-                    }
-                }
-
-                if (outdated) {
-                    setOutdatedVersionFlag(true);
-                } else {
-                    console.log('Aircraft version ok');
-                }
-
-                setVersionChecked(true);
-            } catch (error) {
-                console.error('Version comparison failed: ', error);
-            }
+            const done:boolean = AircraftVersionChecker.checkVersion(buildInfo, releaseInfo, newestCommit, newestExpCommit);
+            setVersionChecked(done);
         }
     }, [buildInfo, releaseInfo, newestCommit, newestExpCommit]);
